chore: optimized map entry listener
