+++ horizon-core/src/main/java/de/telekom/eni/pandora/horizon/cache/listener/AbstractHazelcastJsonEvent.java
@@ -1,4 +1,4 @@
-package de.telekom.eni.pandora.horizon.cache.service;
+package de.telekom.eni.pandora.horizon.cache.listener;
 
 import com.hazelcast.core.EntryEvent;
 import com.hazelcast.core.HazelcastJsonValue;
@@ -9,8 +9,8 @@ import lombok.Setter;
 @Getter
 @Setter
 @AllArgsConstructor
-public class JsonCacheServiceEvent<T> {
-    private T newValue;
+public abstract class AbstractHazelcastJsonEvent<T> {
+    private T value;
 
     private T oldValue;
